pyproject.toml
[tool.poetry]
name = "holy-hack-2024-ai-workshop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
streamlit = "^1.31.1"
st-pages = "^0.4.5"
ipykernel = "^6.29.3"
jupyter = "^1.0.0"
librosa = "^0.10.1"
matplotlib = "^3.8.3"
pydub = "^0.25.1"
langchain = "^0.1.11"
qdrant-client = "^1.7.3"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
langchain-experimental = "^0.0.53"
semantic-router = "^0.0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
